+++ src/components/Projects.js
@@ -3,34 +3,27 @@ import Table from "react-bootstrap/Table";
 import { useState } from "react";
 import { Button } from "@mui/material";
 import "bootstrap/dist/css/bootstrap.min.css";
-//list of projects is generated using the prop received
-const ProjectList = (props) => {
+
+const Projects = (props) => {
   const [visible, setVisible] = useState(10);
   const data = props.items;
   const database = props.repository;
-  const directWord = props.directWord;
-  const  word = props.word;
   var tag;
-  var dummy =[]
 
-  //console.log(database);
   let Visible = visible;
 
   const loadMOre = () => {
     setVisible(Visible + 10);
   };
 
-
-
   return (
     <div className="search-items">
       <div className="list">
-        {/* {data.length === 0 && directWord !== "" &&(
+        {data.length === 0 && props.directWord !== "" && (
           <div className="alert-spinner">No results to display...</div>
-        )} */}
+        )}
         {data.slice(0, visible).map((item) => {
           return (
-            /* all columns set to 20% total width */
             <div className="preview" key={item.id}>
               <Table className="table">
                 <thead>
@@ -84,16 +77,28 @@ const ProjectList = (props) => {
                   </tr>
                   <tr className="attributes">
                     <td className="watchers">
-                      {database === "GitHub" && <h1 className="headers">
-                        watchers : {item.watchers_count}
-                      </h1>}
-                      {database === "MATC" && <h1 className="headers">
-                        downloads : {item.downloads}
-                      </h1>}
-                      {database === "All" && <h1 className="headers">
-                        {tag = (item.homepage_url === null || item.homepage_url === "") ? "watchers" : "downloads"} : {item.downloads || item.watchers_count}
-                      </h1>}                      
-              
+                      {database === "GitHub" && (
+                        <h1 className="headers">
+                          watchers : {item.watchers_count}
+                        </h1>
+                      )}
+                      {database === "MATC" && (
+                        <h1 className="headers">
+                          downloads : {item.downloads}
+                        </h1>
+                      )}
+                      {database === "All" && (
+                        <h1 className="headers">
+                          {
+                            (tag =
+                              item.homepage_url === null ||
+                              item.homepage_url === ""
+                                ? "watchers"
+                                : "downloads")
+                          }{" "}
+                          : {item.downloads || item.watchers_count}
+                        </h1>
+                      )}
                     </td>
                     <td className="language">
                       <h1 className="headers">
@@ -101,30 +106,56 @@ const ProjectList = (props) => {
                       </h1>
                     </td>
                     <td className="forks">
-                      {database === "GitHub" && <h1 className="headers">
-                        forks : {item.forks_count}
-                      </h1>}
-                      {database === "MATC" && <h1 className="headers">
-                        comments : {item.no_of_comments}
-                      </h1>}
-                      {database === "All" && <h1 className="headers">
-                        {tag = (item.homepage_url === null || item.homepage_url === "") ? "forks" : "comments"} : {item.downloads || item.watchers_count}
-                      </h1>} 
-                      
+                      {database === "GitHub" && (
+                        <h1 className="headers">forks : {item.forks_count}</h1>
+                      )}
+                      {database === "MATC" && (
+                        <h1 className="headers">
+                          comments : {item.no_of_comments}
+                        </h1>
+                      )}
+                      {database === "All" && (
+                        <h1 className="headers">
+                          {
+                            (tag =
+                              item.homepage_url === null ||
+                              item.homepage_url === ""
+                                ? "forks"
+                                : "comments")
+                          }{" "}
+                          : {item.downloads || item.watchers_count}
+                        </h1>
+                      )}
                     </td>
                     <td className="license">
-                      <h1 className="headers">license : {(item.license.length < 50 && item.license) || "BDS"}</h1>
+                      <h1 className="headers">
+                        license :{" "}
+                        {(item.license.length < 50 && item.license) || "BDS"}
+                      </h1>
                     </td>
                     <td>
-                    {database === "GitHub" && <h1 className="headers">
-                        stargazers : {item.stargazers_count || "0"}
-                      </h1>}
-                      {database === "MATC" && <h1 className="headers">
-                        ratings : {item.no_of_ratings || "0"}
-                      </h1>}
-                      {database === "All" && <h1 className="headers">
-                        {tag = (item.homepage_url === null || item.homepage_url === "") ? "stargazers" : "ratings"} : {item.stargazers_count || item.no_of_ratings}
-                      </h1>} 
+                      {database === "GitHub" && (
+                        <h1 className="headers">
+                          stargazers : {item.stargazers_count || "0"}
+                        </h1>
+                      )}
+                      {database === "MATC" && (
+                        <h1 className="headers">
+                          ratings : {item.no_of_ratings || "0"}
+                        </h1>
+                      )}
+                      {database === "All" && (
+                        <h1 className="headers">
+                          {
+                            (tag =
+                              item.homepage_url === null ||
+                              item.homepage_url === ""
+                                ? "stargazers"
+                                : "ratings")
+                          }{" "}
+                          : {item.stargazers_count || item.no_of_ratings}
+                        </h1>
+                      )}
                     </td>
                   </tr>
                 </tbody>
@@ -133,7 +164,7 @@ const ProjectList = (props) => {
           );
         })}
         <div className="page-count">
-          {/* {visible < data.length && (
+          {visible < data.length && (
             <Button
               className="loader"
               variant="contained"
@@ -142,11 +173,11 @@ const ProjectList = (props) => {
             >
               Next 10
             </Button>
-          )} */}
+          )}
         </div>
       </div>
     </div>
   );
 };
 
-export default ProjectList;
+export default Projects;
