+++ src/index.js
@@ -4,9 +4,9 @@ import './index.css';
 import App from './App';
 //import reportWebVitals from './reportWebVitals';
 import { Amplify } from 'aws-amplify';
-//import awsExports from './aws-exports';
+import awsExports from './aws-exports';
 import "react-datepicker/dist/react-datepicker.css";
-//Amplify.configure(awsExports);
+Amplify.configure(awsExports);
 
 ReactDOM.render(
   <React.StrictMode>
